requirements.txt
balldontlie==0.1.6
geopy==2.4.1
joblib==1.4.2
numpy==1.24.3
pandas==2.0.3
psycopg2==2.9.10
python-dotenv==1.1.0
scikit_learn==1.3.0
SQLAlchemy==1.4.39
streamlit==1.44.0
xgboost==3.0.0
